backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
yara-python==4.3.1
pypdf2==3.0.1
reportlab==4.0.7
pydantic==2.5.0
python-json-logger==2.0.7

